+++ SamsungPay/src/main/java/com/braintreepayments/api/SamsungPayInternalClient.java
@@ -151,7 +151,7 @@ class SamsungPayInternalClient {
             @Override
             public void onFailure(int errorCode, Bundle bundle) {
                 if (errorCode == SpaySdk.ERROR_USER_CANCELED) {
-                    UserCanceledException userCanceledError = new UserCanceledException("User canceled Samsung Pay.");
+                    UserCanceledException userCanceledError = new UserCanceledException("User canceled Samsung Pay.", true);
                     listener.onSamsungPayStartError(userCanceledError);
                 } else {
                     SamsungPayException samsungPayError = new SamsungPayException(errorCode);